Permissions, Logging, Admin page work, Error handling, build fixes, use yarn workspaces, add constants (#75)
* add todo

* feat(constants): add constants package

* feat(app): lots of work

couple weeks worth of work without commits. whoops. Added authentication, error handling, fixed lots
of bugs, autologin on web-client, lots of new admin functionality

* use yarn workspaces

* revertme: deploy api

* nohoist?

* debug

* feat(travis): upgrade yarn to latest

* attempt

* remove procfile

* add Procfile

* yarn start

* attempt at builting

* use latest node

* typeo

* update pwd before running api

* fix procfile path

* parallelize build, restore dev branch deploy

* remove nohoist

* fix typo in package.json

* docs(api/RouterMiddlewareFactory): update docs regarding usage of request-scoped di container

* fix(api/RouterMiddlewareFactory): remove dead code

* feat(api/RouterMiddlewareFactory): add explicit error checking for ctx.state.requestContainer

update docs

* feat(api/MethodNotSupportedError): add allow property

* test(api/RouteRegistry): refactor tests

refactor tests to expect thrown errors instead of error metadata object

* feat(api/RouteRegistry): throw instead of return errors

* fix(api/RestRepositoryController): update usage of routeRegistry#lookupRoute

* refactor(api/AuthorizationMiddlewareFactory): refactor usage of lookupRoute

* refactor(api): refactor for new error flow from lookupRoute

move error handling for OPTIONS/ALLOW: headers up to ErrorHandlingMiddleware. Improve
errorhandlingmiddleware tests.

* fix(api/UserController): fix permissions

* refactor(api): s/requiredRoles/allowedRoles/g

* fix(api/Route.ts): make default allowed role consistent, and set to "public"

* fix(api/controllers): fix PrizeController permissions

remove unused controllers

* docs(api/MethodNotSupportedError): clarify usage

* fix(web-client/HalloweenApp): remove console log

* chore(web-client/AuthenticationService): lint

* fix(web-client/HalloweenApp): improve error reporting

* fix(web-client/AdminPage): remove dead code

clean up weird stuff

* feat(web-client): improve api error handling and auth workflow

* fix(api/UserController.ts): fix default permissions

* fix(api/tsconfig): remove noimplicitusestrict

* feat(web-client/AdminUsersTab): reimplement using handlerFactory memoized currying helper

* feat(api/tsconfig): enable noUnusedLocals error

e

* feat(web-client/tsconfig): enable noUnusedLocals error

* feat(api/PrizeController): fix permissions

* feat(api/RouteRegistry): rename "handler" to "resolver" in IRouteHandler

* docs(constants/README): remove junk

* docs(constants/package.json): remove some info

* refactor(web-client/AdminPage): make use of member declaration syntax

* refactor(web-client/AdminPrizeTab): move spinner control up to AdminPage

* docs(api/AuthenticationError): add better docs

* test(web-client): add mock roles

* refactor(api/AuthenticationService):

* fix(api/OrmContext): wait to create initial entites until all repositories are loaded

* refactor(api/ICodedException.ts): rename to ICodedError

* refactor(api/db): change out "exception" for "error" nomenclature

* fix(api/logging): set jest log level to error

* docs(AuthorizationMiddlewareFactoryr): clarify container population work done by middleware

* docs(api/middlewares/README): update docs

* docs(rendermiddlewarefactory): improve docs

+++ packages/api/src/controllers/RestRepositoryController.ts
@@ -1,8 +1,15 @@
 import { Context } from "koa";
-import { Connection, DeleteResult, Repository } from "typeorm";
+import { Connection, Repository } from "typeorm";
+import { RoleLiteral } from "../auth/RoleLiteral";
 import { asClassMethod } from "../AwilixHelpers";
 import { HttpMethod } from "../HttpMethod";
+import { MethodNotSupportedError } from "../web/MethodNotSupportedError";
 import { RouteRegistry } from "../web/RouteRegistry";
+import { UnknownRouteError } from "../web/UnknownRouteError";
+/**
+ * Pluralize a given string.
+ * @param str - The string to pluralize.
+ */
 const pluralize: (str: string) => string = (str: string) => {
   if (str.endsWith("s")) {
     return `${str}es`;
@@ -12,8 +19,24 @@ const pluralize: (str: string) => string = (str: string) => {
   }
   return `${str}s`;
 };
+
+/**
+ * Creates a baseroute for the given class.
+ * @param clazz - The entity class for the route.
+ */
 export function getRoute(clazz: new () => any) {
   return "/" + clazz.name[0].toLowerCase() + clazz.name.substr(1);
+}
+/**
+ * Interface for objects mapping HTTP methods and routes to class methods.
+ */
+export interface IFallbackHandlerMap {
+  [key: string]: {
+    [method in HttpMethod]?: {
+      fn: (...args: any[]) => any;
+      roles: RoleLiteral[];
+    }
+  };
 }
 export abstract class RestRepositoryController<T> {
   protected modelClass: new () => T;
@@ -21,6 +44,7 @@ export abstract class RestRepositoryController<T> {
   protected baseRoute: string;
   protected listRoute: string;
   protected detailRoute: string;
+  protected abstract defaultRoles: RoleLiteral[];
   constructor(orm: Connection, modelClass: new () => T) {
     this.modelClass = modelClass;
     this.repository = orm.getRepository(this.modelClass);
@@ -28,37 +52,76 @@ export abstract class RestRepositoryController<T> {
     this.listRoute = pluralize(this.baseRoute);
     this.detailRoute = `${this.listRoute}/{id}`;
   }
+
   /**
    * Registers default fallback handlers for this repository if they are
    * not already registered.
    * @param routeRegistry The route registry to write to.
    */
   registerRoutes(routeRegistry: RouteRegistry) {
-    interface IFallbackHandlerMap {
-      [key: string]: { [method in HttpMethod]?: (...args: any[]) => any };
-    }
     const fallbackHandlers: IFallbackHandlerMap = {
       [this.listRoute]: {
-        [HttpMethod.GET]: this.getAll,
-        [HttpMethod.POST]: this.createOne
+        [HttpMethod.GET]: {
+          fn: this.getAll,
+          roles: this.defaultRoles
+        },
+        [HttpMethod.POST]: {
+          fn: this.createOne,
+          roles: this.defaultRoles
+        }
       },
       [this.detailRoute]: {
-        [HttpMethod.GET]: this.getOne,
-        [HttpMethod.DELETE]: this.deleteOne,
-        [HttpMethod.PATCH]: this.updateOne
+        [HttpMethod.GET]: { fn: this.getOne, roles: this.defaultRoles },
+        [HttpMethod.DELETE]: { fn: this.deleteOne, roles: this.defaultRoles },
+        [HttpMethod.PATCH]: { fn: this.updateOne, roles: this.defaultRoles }
       }
     };
+    // allow for configuration of the map before applying it
+    this.configureFallbackHandlers(fallbackHandlers);
+
+    // apply the map without overwriting any existing routes
     for (const route of Object.keys(fallbackHandlers)) {
       const methodMap = fallbackHandlers[route];
       for (const method of Object.keys(methodMap) as HttpMethod[]) {
-        const { error } = routeRegistry.lookupRoute(route, method);
-        if (error) {
-          // handler isn't covered, register the route
-          const resolver = asClassMethod(this, methodMap[method]);
-          routeRegistry.registerRoute(route, method, resolver);
+        try {
+          routeRegistry.lookupRoute(route, method);
+        } catch (e) {
+          if (
+            e instanceof UnknownRouteError ||
+            e instanceof MethodNotSupportedError
+          ) {
+            // handler isn't covered, register it
+            const resolver = asClassMethod(this, methodMap[method].fn);
+            routeRegistry.registerRoute(
+              route,
+              method,
+              resolver,
+              methodMap[method].roles
+            );
+          } else {
+            throw e;
+          }
         }
       }
     }
+  }
+
+  /**
+   * Overridable method that allows for the fallback route handler map to
+   * be modified prior to being applied. Default implementation is a no-op.
+   *
+   * This method will be invoked with the fallback handler map that is used to
+   * generate default endpoints for the controller.
+   *
+   * This method may (should) modify the map in-place.
+   *
+   * @param fallbackHandlers - A route map. This map will be iterated over
+   * and any (route, method) pairs defined in it that are not already registered
+   * will be registered.
+   * @see registerRoutes
+   */
+  configureFallbackHandlers(fallbackHandlers: IFallbackHandlerMap): void {
+    // noop
   }
 
   /**
@@ -95,7 +158,7 @@ export abstract class RestRepositoryController<T> {
    */
   async updateOne(id: string, requestBody: any, ctx: Context) {
     try {
-      const result = await this.repository.update(id, requestBody);
+      await this.repository.update(id, requestBody);
       return await this.getOne(id);
     } catch (error) {
       ctx.status = 400;
